requirements.txt
streamlit==1.42.0
google-generativeai==0.8.4
PyPDF2==3.0.1
python-docx==1.1.2
